api/requirements.txt
# argo version with 3 added to major version (argo 3.4.8 == argo-workflows 6.4.8) https://github.com/argoproj/argo-workflows/tree/master/sdks/python#versioning
argo-workflows==6.4.8
pyyaml
requests
fastapi
fastapi-users[beanie]
google-cloud-bigquery
uvicorn[standard]
httpx_oauth==0.15.1
minio
pydantic-settings
google-cloud-logging
pyproj
wget
shapely
pyproj
pandas
